Refactor on views file
