+++ unifier/apps/drf/v1/views/__init__.py
@@ -1,0 +1,2 @@
+from unifier.apps.drf.v1.views.manga import MangaChapterRetrieveViewSet, MangaViewSet
+from unifier.apps.drf.v1.views.novel import NovelChapterRetrieveViewSet, NovelViewSet
